Move binary_to_label() to "labelimage" module
