+++ src/imcflibs/imagej/labelimage.py
@@ -6,6 +6,7 @@ from ij.plugin.filter import ImageProcessor, ThresholdToSelection
 from ij.process import FloatProcessor, ImageProcessor
 from inra.ijpb.label import LabelImages as li
 from inra.ijpb.plugins import AnalyzeRegions
+from mcib3d.image3d import ImageHandler, ImageLabeller
 
 
 def label_image_to_roi_list(label_image, low_thresh=None):
@@ -138,3 +139,44 @@ def measure_objects_size_shape_2d(label_image):
     """
     regions = AnalyzeRegions()
     return regions.process(label_image)
+
+
+def binary_to_label(imp, title, min_thresh=1, min_vol=None, max_vol=None):
+    """Segment a binary image to get a label image (2D/3D).
+
+    Works on: 2D and 3D binary data.
+
+    Parameters
+    ----------
+    imp : ImagePlus
+        Binary 3D stack or 2D image.
+    title : str
+        Title of the new image.
+    min_thresh : int, optional
+        Threshold to do segmentation, also allows for label filtering, by
+        default 1.
+    min_vol : float, optional
+        Volume under which to exclude objects, by default None.
+    max_vol : float, optional
+        Volume above which to exclude objects, by default None.
+
+    Returns
+    -------
+    ImagePlus
+        Segmented labeled ImagePlus.
+    """
+    cal = imp.getCalibration()
+    img = ImageHandler.wrap(imp)
+    img = img.threshold(min_thresh, False, False)
+
+    labeler = ImageLabeller()
+    if min_vol:
+        labeler.setMinSize(min_vol)
+    if max_vol:
+        labeler.setMaxSize(max_vol)
+
+    seg = labeler.getLabels(img)
+    seg.setScale(cal.pixelWidth, cal.pixelDepth, cal.getUnits())
+    seg.setTitle(title)
+
+    return seg.getImagePlus()
